Replace most refs to Microsoft with Windows
Referring to software as "Microsoft software" implies it's written
by Microsoft, which I don't think the Mares Dive Organizer is.
Microsoft still needs to stay for SQL Compact Edition which is theirs
but I think "Windows software" is the better description for the rest.
For consistency I think this works for refering to the OS too, as we
use Mac, not Apple, and Linux not Canonical.

Signed-off-by: Tim Wootton <tim@tee-jay.demon.co.uk>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>