User Manual: Minor updates to text with 4 image replacements
Minor grammatical and stylistic text adjustsments, following quite a
few changes to the text by several contributors over the last two weeks.

Four images are replaced to reflect the latest UI, and also to
reflect the Gnome 3 rendering. These relate to
the specification of cylinder gas in the Equipment Tab.

Signed-off-by: Willem Ferguson <willemferguson@zoology.up.ac.za>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>