CCR patch: Adapt pressure interpolation for CCR (2)
This is the second pathch in this series of four, allowing the calculation
of cylinder pressures for CCR equipment.

Change function fill_missing_tank_pressures in order to enable working
with the diuent gas, comprising a complete set of pressures kept separate
from all other tank pressures.
Flag diluent_flag indicates calculations for the diluent cylinder.

[Dirk Hohndel: serious whitespace cleanup]

Signed-off-by: willem ferguson <willemferguson@zoology.up.ac.za>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>