Fix dc nickname helper function
This helper function is used to get a nickname for a dive computer, when
the model information on its own may be ambiguous (ie there may be
multiple dive computers of the same model, and we've nicknamed them by
owner).

However, the helper did completely the wrong thing if it didn't find a
dive computer entry at all due to a missing device ID - it would just
return empty.  Which is bogus: it should return the model name, the same
way it does if the nickname is missing.

Signed-off-by: Linus Torvalds <torvalds@linux-foundation.org>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>