Replace Poseidon example dive log files in dives directory
The present example files have dives with some unusual
characteristics due to the fact that the dive pushes the
limits of the Poseidon recreational rebreather.
Replace these example files with a dive that does not have
any unusual characteristics. This facilitates debugging of the
Poseidon code. The xml file represents the same dive as the two
text files. The xml file will ultimately be removed.

Signed-off-by: willem ferguson <willemferguson@zoology.up.ac.za>
Signed-off-by: Dirk Hohndel <dirk@hohndel.org>